backend/requirements.txt

Flask>=2.0.1
Flask-Cors>=3.0.10
Werkzeug>=2.0.1
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.3.1
requests>=2.26.0
python-dotenv>=0.19.1
psycopg2-binary>=2.9.3
SQLAlchemy>=1.4.40
supabase>=1.0.3
gunicorn>=20.1.0
pytest>=7.0.0
flask-swagger-ui>=4.11.1
tensorflow>=2.6.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
coverage>=6.2
pyyaml>=6.0
PyJWT>=2.6.0
json-logging>=1.3.0
celery>=5.2.7
redis>=4.5.1
prometheus-client>=0.16.0
flask-limiter>=3.3.0
apscheduler>=3.10.1
cryptography>=38.0.4
pydantic>=1.10.4
flask-pydantic>=0.11.0
python-jose>=3.3.0
passlib>=1.7.4
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pylint>=2.15.10
black>=23.1.0
pytest-mock>=3.10.0
flask-httpauth>=4.7.0
connexion>=2.14.1
marshmallow>=3.19.0
captum>=0.6.0
sentry-sdk>=1.19.1
# Commenting out conflicting GraphQL dependencies for now
# graphene>=3.2.1
# flask-graphql>=2.0.1
# graphql-core>=3.2.3
# graphql-relay>=3.2.0
psutil>=5.9.4
kubernetes>=26.1.0
imgaug>=0.4.0
albumentations>=1.3.0
shap>=0.41.0
tensorboard>=2.6.0
opencv-contrib-python>=4.5.0
tensorflow-addons>=0.14.0
tensorflow-hub>=0.12.0
lime>=0.2.0.1
efficientnet>=1.1.1
gdown>=4.6.0
# New dependencies for advanced features
librosa>=0.9.2  # For audio feature extraction
mtcnn>=0.1.1  # For face detection and alignment
# dlib>=19.0.0  # For facial landmark detection
onnx>=1.13.0  # For ONNX model export
tf2onnx>=1.13.0  # For TensorFlow to ONNX conversion
onnxruntime>=1.13.1  # For ONNX model inference
optuna>=3.1.0  # For hyperparameter tuning
tensorflow-model-optimization>=0.7.3  # For model quantization
transformers>=4.26.0  # For Vision Transformer components
tensorflow-probability>=0.14.0  # For probabilistic models
tqdm>=4.64.1  # For progress bars
pytest-xdist>=3.2.1  # For parallel testing

# PDF report generation
reportlab>=3.6.12  # For PDF generation
PyPDF2>=3.0.0  # For PDF manipulation
pdfkit>=1.0.0  # For HTML to PDF conversion
weasyprint>=58.1  # Alternative HTML to PDF converter

# Enhanced skin tone analysis
scikit-image>=0.19.3  # For advanced image processing
face-recognition>=1.3.0  # For face recognition and analysis
opencv-python-headless>=4.7.0  # Headless OpenCV for servers
soundfile>=0.12.1  # For audio file handling
scipy>=1.10.1  # For scientific computing
